+++ src/pages/Diary/components/DiaryCommon/DiaryHeader.tsx
@@ -1,0 +1,12 @@
+const DiaryHeader = () => {
+  return (
+    <>
+      <div className="font-title flex gap-2 font-bold">
+        <button className="text-grey-300">{'<'}</button>
+        <span>김선익 내과</span>
+      </div>
+    </>
+  );
+};
+
+export default DiaryHeader;
