+++ src/pages/Diary/components/DiaryCommon/index.ts
@@ -1,1 +1,2 @@
 export { default as DiaryCategoryList } from './DiaryCategoryList';
+export { default as DiaryHeader } from './DiaryHeader';
